Add optional limit to versions API
So that clients can query more than the top 5 Ruby versions and top 3
platforms, add a new, optional limit parameter which defaults to the
previous values but allows clients to specify something larger.

* /versions/ruby.json will still return the top 5 Ruby versions with
  all others grouped under "Other"
* /versions/ruby.json?limit=10 will return the top 10 Ruby versions with
  all others grouped under "Other"

We could send _all_ the versions to the frontend and filter them there
but, this way, we don't increase the size of the current API responses
and allow other API clients to query data in more detail.